pyproject.toml
[project]
name = "project-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.24",
]
